+++ src/main/java/me/spoony/botanico/common/buildings/BuildingGrass.java
@@ -27,7 +27,7 @@ public class BuildingGrass extends Building {
   }
 
   @Override
-  public void render(RendererGame rg, ClientPlane level, OmniPosition position, byte d,
+  public void render(RendererGame rg, ClientPlane level, OmniPosition position, int d,
       Color color) {
     int val = hash(position.getTileX(), position.getTileY()) % 4;
     rg.sprite(position,
